+++ tests/functional_tests/ffront_tests/test_program.py
@@ -58,151 +58,6 @@ def test_identity_fo_execution(identity_def):
     assert np.allclose(in_field, out_field)
 
 
-def test_copy_parsing(copy_program_def):
-    past_node = ProgramParser.apply_to_function(copy_program_def)
-
-    field_type = common_types.FieldType(
-        dims=[IDim],
-        dtype=common_types.ScalarType(kind=common_types.ScalarKind.FLOAT64, shape=None),
-    )
-    pattern_node = P(
-        past.Program,
-        id=eve.SymbolName("copy_program"),
-        params=[
-            P(past.Symbol, id=eve.SymbolName("in_field"), type=field_type),
-            P(past.Symbol, id=eve.SymbolName("out_field"), type=field_type),
-        ],
-        body=[
-            P(
-                past.Call,
-                func=P(past.Name, id=past.SymbolRef("identity")),
-                args=[P(past.Name, id=past.SymbolRef("in_field"))],
-                kwargs={"out": P(past.Name, id=past.SymbolRef("out_field"))},
-            )
-        ],
-        location=P(past.SourceLocation),
-    )
-    assert pattern_node.match(past_node, raise_exception=True)
-
-
-def test_double_copy_parsing(double_copy_program_def):
-    past_node = ProgramParser.apply_to_function(double_copy_program_def)
-
-    field_type = common_types.FieldType(
-        dims=[IDim],
-        dtype=common_types.ScalarType(kind=common_types.ScalarKind.FLOAT64, shape=None),
-    )
-    pattern_node = P(
-        past.Program,
-        id=eve.SymbolName("double_copy_program"),
-        params=[
-            P(past.Symbol, id=eve.SymbolName("in_field"), type=field_type),
-            P(past.Symbol, id=eve.SymbolName("intermediate_field"), type=field_type),
-            P(past.Symbol, id=eve.SymbolName("out_field"), type=field_type),
-        ],
-        body=[
-            P(
-                past.Call,
-                func=P(past.Name, id=past.SymbolRef("identity")),
-                args=[P(past.Name, id=past.SymbolRef("in_field"))],
-                kwargs={"out": P(past.Name, id=past.SymbolRef("intermediate_field"))},
-            ),
-            P(
-                past.Call,
-                func=P(past.Name, id=past.SymbolRef("identity")),
-                args=[P(past.Name, id=past.SymbolRef("intermediate_field"))],
-                kwargs={"out": P(past.Name, id=past.SymbolRef("out_field"))},
-            ),
-        ],
-    )
-    assert pattern_node.match(past_node, raise_exception=True)
-
-
-def test_undefined_field_program(identity_def):
-    identity = field_operator(identity_def)
-
-    def undefined_field_program(in_field: Field[[IDim], "float64"]):
-        identity(in_field, out=out_field)
-
-    with pytest.raises(
-        ProgramTypeError,
-        match=(r"Undeclared or untyped symbol `out_field`."),
-    ):
-        ProgramParser.apply_to_function(undefined_field_program)
-
-
-@pytest.mark.xfail
-def test_inout_prohibited(identity_def):
-    identity = field_operator(identity_def)
-
-    def inout_field_program(inout_field: Field[[IDim], "float64"]):
-        identity(inout_field, out=inout_field)
-
-    with pytest.raises(
-        GTTypeError,
-        match=(r"Call to function with field as input and output not allowed."),
-    ):
-        ProgramLowering.apply(ProgramParser.apply_to_function(inout_field_program))
-
-
-def test_invalid_call_sig_program(invalid_call_sig_program_def):
-    with pytest.raises(
-        GTTypeError,
-    ) as exc_info:
-        ProgramLowering.apply(ProgramParser.apply_to_function(invalid_call_sig_program_def))
-
-    assert exc_info.match("Invalid call to `identity`")
-    # TODO(tehrengruber): find a better way to test this
-    assert (
-        re.search(
-            "Function takes 1 arguments, but 2 were given.", exc_info.value.__context__.args[0]
-        )
-        is not None
-    )
-    assert (
-        re.search(
-            "Missing required keyword argument\(s\) `out`", exc_info.value.__context__.args[0]
-        )
-        is not None
-    )
-
-
-def test_copy_restrict_parsing(copy_restrict_program_def):
-    past_node = ProgramParser.apply_to_function(copy_restrict_program_def)
-
-    field_type = common_types.FieldType(
-        dims=[IDim],
-        dtype=common_types.ScalarType(kind=common_types.ScalarKind.FLOAT64, shape=None),
-    )
-    slice_pattern_node = P(
-        past.Slice, lower=P(past.Constant, value=1), upper=P(past.Constant, value=2)
-    )
-    pattern_node = P(
-        past.Program,
-        id=eve.SymbolName("copy_restrict_program"),
-        params=[
-            P(past.Symbol, id=eve.SymbolName("in_field"), type=field_type),
-            P(past.Symbol, id=eve.SymbolName("out_field"), type=field_type),
-        ],
-        body=[
-            P(
-                past.Call,
-                func=P(past.Name, id=past.SymbolRef("identity")),
-                args=[P(past.Name, id=past.SymbolRef("in_field"))],
-                kwargs={
-                    "out": P(
-                        past.Subscript,
-                        value=P(past.Name, id=past.SymbolRef("out_field")),
-                        slice_=slice_pattern_node,
-                    )
-                },
-            )
-        ],
-    )
-
-    pattern_node.match(past_node, raise_exception=True)
-
-
 def test_shift_by_one_execution():
     size = 10
     in_field = np_as_located_field(IDim)(np.arange(0, size, 1))
